Several changes & improvements
- Rename 'Time Granularity' to 'Tick Size'

Came to reframe what we're doing as really measuring "ticks since epoch"
for the time component, and we're just controlling the size of this
tick. It's also shorter and better for the API.

- Add ability to provide own time and randomness sources

Makes it possible to deterministically test functionality related to
STIDs. We add a test which does this.

- Remove DefaultConfig()

NewConfig() does the same thing.

- Rename 'RandomChars' to 'NumRandomChars'

A bit clearer we're talking about an integer. Also conflicts a little
less with 'WithRandomSource' now that this exists.

- Validate alphabet doesn't contain duplicate chars

On generator creation.