Add stid -> flexid migration docs
Not ideal, but by doing it this way, we avoid the need to forever do
`/v2` installations, which would be a shame to require all because of a
rename so early on in the project's lifetime. Pay a little cost now to
save more in the future, is the idea.